+++ cvat-ui/src/components/annotation-page/standard-workspace/canvas-wrapper.tsx
@@ -58,13 +58,13 @@ interface Props {
     contrastLevel: number;
     saturationLevel: number;
     resetZoom: boolean;
-    contextVisible: boolean;
-    contextType: ContextMenuType;
     aamZoomMargin: number;
     showObjectsTextAlways: boolean;
+    showAllInterpolationTracks: boolean;
     workspace: Workspace;
     automaticBordering: boolean;
     keyMap: Record<string, ExtendedKeyMapOptions>;
+    canvasBackgroundColor: string;
     switchableAutomaticBordering: boolean;
     onSetupCanvas: () => void;
     onDragCanvas: (enabled: boolean) => void;
@@ -93,6 +93,7 @@ interface Props {
     onChangeGridColor(color: GridColor): void;
     onSwitchGrid(enabled: boolean): void;
     onSwitchAutomaticBordering(enabled: boolean): void;
+    onFetchAnnotation(): void;
 }
 
 export default class CanvasWrapperComponent extends React.PureComponent<Props> {
@@ -137,6 +138,7 @@ export default class CanvasWrapperComponent extends React.PureComponent<Props> {
             curZLayer,
             resetZoom,
             grid,
+            gridSize,
             gridOpacity,
             gridColor,
             brightnessLevel,
@@ -145,8 +147,11 @@ export default class CanvasWrapperComponent extends React.PureComponent<Props> {
             workspace,
             frameFetching,
             showObjectsTextAlways,
+            showAllInterpolationTracks,
             automaticBordering,
             showProjections,
+            canvasBackgroundColor,
+            onFetchAnnotation,
         } = this.props;
 
         if (prevProps.showObjectsTextAlways !== showObjectsTextAlways
@@ -159,6 +164,10 @@ export default class CanvasWrapperComponent extends React.PureComponent<Props> {
                 autoborders: automaticBordering,
                 showProjections,
             });
+        }
+
+        if (prevProps.showAllInterpolationTracks !== showAllInterpolationTracks) {
+            onFetchAnnotation();
         }
 
         if (prevProps.sidebarCollapsed !== sidebarCollapsed) {
@@ -177,6 +186,10 @@ export default class CanvasWrapperComponent extends React.PureComponent<Props> {
             if (el) {
                 (el as any).instance.fill({ opacity: opacity / 100 });
             }
+        }
+
+        if (gridSize !== prevProps.gridSize) {
+            canvasInstance.grid(gridSize, gridSize);
         }
 
         if (gridOpacity !== prevProps.gridOpacity
@@ -222,7 +235,8 @@ export default class CanvasWrapperComponent extends React.PureComponent<Props> {
         }
 
         if (prevProps.opacity !== opacity || prevProps.blackBorders !== blackBorders
-            || prevProps.selectedOpacity !== selectedOpacity || prevProps.colorBy !== colorBy) {
+            || prevProps.selectedOpacity !== selectedOpacity || prevProps.colorBy !== colorBy
+        ) {
             this.updateShapesView();
         }
 
@@ -240,6 +254,13 @@ export default class CanvasWrapperComponent extends React.PureComponent<Props> {
                 loadingAnimation.classList.remove('cvat_canvas_hidden');
             } else {
                 loadingAnimation.classList.add('cvat_canvas_hidden');
+            }
+        }
+
+        if (prevProps.canvasBackgroundColor !== canvasBackgroundColor) {
+            const canvasWrapperElement = window.document.getElementsByClassName('cvat-canvas-container').item(0) as HTMLElement | null;
+            if (canvasWrapperElement) {
+                canvasWrapperElement.style.backgroundColor = canvasBackgroundColor;
             }
         }
 
@@ -382,10 +403,9 @@ export default class CanvasWrapperComponent extends React.PureComponent<Props> {
         const {
             activatedStateID,
             onUpdateContextMenu,
-            contextType,
         } = this.props;
 
-        if (contextType !== ContextMenuType.CANVAS_SHAPE_POINT) {
+        if (e.target && !(e.target as HTMLElement).classList.contains('svg_select_points')) {
             onUpdateContextMenu(activatedStateID !== null, e.clientX, e.clientY,
                 ContextMenuType.CANVAS_SHAPE);
         }
@@ -618,14 +638,12 @@ export default class CanvasWrapperComponent extends React.PureComponent<Props> {
         const {
             annotations,
             frameData,
-            frameAngle,
             canvasInstance,
         } = this.props;
 
         if (frameData !== null) {
             canvasInstance.setup(frameData, annotations
                 .filter((e) => e.objectType !== ObjectType.TAG));
-            canvasInstance.rotate(frameAngle);
         }
     }
 
@@ -639,6 +657,7 @@ export default class CanvasWrapperComponent extends React.PureComponent<Props> {
             brightnessLevel,
             contrastLevel,
             saturationLevel,
+            canvasBackgroundColor,
         } = this.props;
 
         // Size
@@ -663,6 +682,11 @@ export default class CanvasWrapperComponent extends React.PureComponent<Props> {
             backgroundElement.style.filter = `brightness(${brightnessLevel / 100})`
                 + `contrast(${contrastLevel / 100})`
                 + `saturate(${saturationLevel / 100})`;
+        }
+
+        const canvasWrapperElement = window.document.getElementsByClassName('cvat-canvas-container').item(0) as HTMLElement | null;
+        if (canvasWrapperElement) {
+            canvasWrapperElement.style.backgroundColor = canvasBackgroundColor;
         }
 
         // Events