+++ cvat-ui/src/components/annotation-page/standard-workspace/objects-side-bar/objects-side-bar.tsx
@@ -3,7 +3,7 @@
 // SPDX-License-Identifier: MIT
 
 import './styles.scss';
-import React from 'react';
+import React, { useEffect } from 'react';
 import Text from 'antd/lib/typography/Text';
 import Icon from 'antd/lib/icon';
 import Tabs from 'antd/lib/tabs';
@@ -11,11 +11,13 @@ import Layout from 'antd/lib/layout';
 import { RadioChangeEvent } from 'antd/lib/radio';
 import { SliderValue } from 'antd/lib/slider';
 import { CheckboxChangeEvent } from 'antd/lib/checkbox';
+import { Canvas } from 'cvat-canvas-wrapper';
 
 import { ColorBy } from 'reducers/interfaces';
 import ObjectsListContainer from 'containers/annotation-page/standard-workspace/objects-side-bar/objects-list';
 import LabelsListContainer from 'containers/annotation-page/standard-workspace/objects-side-bar/labels-list';
 import AppearanceBlock from './appearance-block';
+
 
 interface Props {
     sidebarCollapsed: boolean;
@@ -26,6 +28,7 @@ interface Props {
     blackBorders: boolean;
     showBitmap: boolean;
     showProjections: boolean;
+    canvasInstance: Canvas;
 
     collapseSidebar(): void;
     collapseAppearance(): void;
@@ -48,6 +51,7 @@ function ObjectsSideBar(props: Props): JSX.Element {
         blackBorders,
         showBitmap,
         showProjections,
+        canvasInstance,
         collapseSidebar,
         collapseAppearance,
         changeShapesColorBy,
@@ -75,6 +79,23 @@ function ObjectsSideBar(props: Props): JSX.Element {
         changeShowBitmap,
         changeShowProjections,
     };
+
+    useEffect(() => {
+        const listener = (event: Event): void => {
+            if ((event as TransitionEvent).propertyName === 'width'
+                    && ((event.target as any).classList as DOMTokenList).contains('ant-tabs-tab-prev')) {
+                canvasInstance.fit();
+            }
+        };
+
+        const [sidebar] = window.document.getElementsByClassName('cvat-objects-sidebar');
+
+        sidebar.addEventListener('transitionstart', listener);
+
+        return () => {
+            sidebar.removeEventListener('transitionstart', listener);
+        };
+    }, []);
 
     return (
         <Layout.Sider