+++ cvat-ui/src/containers/annotation-page/standard-workspace/objects-side-bar/object-item.tsx
@@ -7,8 +7,13 @@ import copy from 'copy-to-clipboard';
 import { connect } from 'react-redux';
 
 import { LogType } from 'cvat-logger';
-import { Canvas, isAbleToChangeFrame } from 'cvat-canvas-wrapper';
-import { ActiveControl, CombinedState, ColorBy } from 'reducers/interfaces';
+import { Canvas } from 'cvat-canvas-wrapper';
+import {
+    ActiveControl,
+    CombinedState,
+    ColorBy,
+    ShapeType,
+} from 'reducers/interfaces';
 import {
     collapseObjectItems,
     changeLabelColorAsync,
@@ -235,6 +240,33 @@ class ObjectItemContainer extends React.PureComponent<Props> {
         copy(url);
     };
 
+    private switchOrientation = (): void => {
+        const { objectState, updateState } = this.props;
+        if (objectState.shapeType === ShapeType.CUBOID) {
+            this.switchCuboidOrientation();
+            return;
+        }
+
+        const reducedPoints = objectState.points.reduce(
+            (acc: number[][], _: number, index: number, array: number[]): number[][] => {
+                if (index % 2) {
+                    acc.push([array[index - 1], array[index]]);
+                }
+
+                return acc;
+            }, [],
+        );
+
+        if (objectState.shapeType === ShapeType.POLYGON) {
+            objectState.points = reducedPoints.slice(0, 1)
+                .concat(reducedPoints.reverse().slice(0, -1)).flat();
+            updateState(objectState);
+        } else if (objectState.shapeType === ShapeType.POLYLINE) {
+            objectState.points = reducedPoints.reverse().flat();
+            updateState(objectState);
+        }
+    };
+
     private toBackground = (): void => {
         const {
             objectState,
@@ -394,7 +426,6 @@ class ObjectItemContainer extends React.PureComponent<Props> {
         this.commit();
     };
 
-
     private switchCuboidOrientation = (): void => {
         function cuboidOrientationIsLeft(points: number[]): boolean {
             return points[12] > points[0];
@@ -444,7 +475,7 @@ class ObjectItemContainer extends React.PureComponent<Props> {
 
     private changeFrame(frame: number): void {
         const { changeFrame, canvasInstance } = this.props;
-        if (isAbleToChangeFrame(canvasInstance)) {
+        if (canvasInstance.isAbleToChangeFrame()) {
             changeFrame(frame);
         }
     }
@@ -534,6 +565,7 @@ class ObjectItemContainer extends React.PureComponent<Props> {
                 copy={this.copy}
                 propagate={this.propagate}
                 createURL={this.createURL}
+                switchOrientation={this.switchOrientation}
                 toBackground={this.toBackground}
                 toForeground={this.toForeground}
                 setOccluded={this.setOccluded}
@@ -552,7 +584,6 @@ class ObjectItemContainer extends React.PureComponent<Props> {
                 changeLabel={this.changeLabel}
                 changeAttribute={this.changeAttribute}
                 collapse={this.collapse}
-                switchCuboidOrientation={this.switchCuboidOrientation}
                 resetCuboidPerspective={() => this.resetCuboidPerspective()}
             />
         );